requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2
SQLAlchemy>=2.0.36
PyJWT==2.8.0
supabase==2.3.4
python-multipart==0.0.9

